+++ src/cupcake/tofu/counting/chain_fusion_samples.py
@@ -88,7 +88,7 @@ def chain_fusion_samples(
             Path(d, fastq_filename) if fastq_filename is not None else None,
         )
 
-    _, count_info = read_count_info(count_filename, dirs, field_to_use)
+    count_info = read_count_info(count_filename, dirs, field_to_use)
 
     # some names may already start with "tmp_" which means they are intermediate results that have already been chained
     # find the first non "tmp_" and start from there